requirements.txt
addict==2.4.0
albumentations==1.3.0
basicsr==1.4.2
blendmodes
einops==0.3.0
gradio==3.33.1
imageio==2.9.0
imageio-ffmpeg==0.4.2
invisible-watermark==0.1.5
kornia==0.6
numba==0.56.4
omegaconf==2.1.1
open_clip_torch==2.0.2
prettytable==3.6.0
pytorch-lightning==1.5.0
safetensors==0.2.7
streamlit==1.12.1
streamlit-drawable-canvas==0.8.0
test-tube==0.7.5
timm==0.6.12
torchmetrics==0.6.0
transformers==4.19.2
webdataset==0.2.5
yapf==0.32.0
xformers==0.0.20
